Changed Various Fields in card.py and set.py to be Optional
For some reason, there are a bunch of fields in card.py and set.py that are not optional. This becomes an issue when using where() with the select keyword, as those fields are required in the select. I don't know how to make tests, so no clue if this change will break anything. I also left the id fields as-is, in case they're necessary for something else.